improvements: доработки кода и изоляция тестов
