+++ backend/src/modules/auth/dto/auth-response.dto.ts
@@ -4,5 +4,6 @@ export class AuthResponseDto {
     id: string;
     email: string;
     name: string;
+    createdAt: Date;
   };
 }
