+++ backend/src/modules/auth/types/auth.types.ts
@@ -1,0 +1,17 @@
+import { UserProfile } from '../domain/user.entity';
+
+export interface RegisterInput {
+  email: string;
+  password: string;
+  name: string;
+}
+
+export interface LoginInput {
+  email: string;
+  password: string;
+}
+
+export interface AuthResult {
+  accessToken: string;
+  user: UserProfile;
+}
